requirements.txt
blinker==1.4
click==6.7
disco==0.5.4
Flask==0.12
flask-reqparse==2.2.1
Flask-Testing==0.6.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
maxminddb==1.3.0
maxminddb-geolite2==2017.208
six==1.10.0
Werkzeug==0.12.1
